clearing App.js to start at a clean slate, moving changes into Modal.js. Downgraded react-scripts due to a WSL bug
